+++ lectures/week05/07.2-swapPointer.cpp
@@ -3,9 +3,9 @@ using std::cin;
 using std::cout;
 using std::endl;
 
-void swap(long *first, long *second) {  // params are pointers
+void swap(int *first, int *second) {  // params are pointers
   // a pointer is a variable, whose value is a memory address
-  long temp;
+  int temp;
   cout << "Value of first is:" << first << " which references " << *first
        << endl;
   temp = *first;  // must derefernce a pointer with *
@@ -14,7 +14,7 @@ void swap(long *first, long *second) {  // params are pointers
 }
 
 int main() {
-  long one = 100, two = 200;
+  int one = 100, two = 200;
   cout << "Pre swap, one has value:" << one << ", two has value:" << two
        << endl;
   swap(&one, &two);  // here & is an operator, address-of
