+++ lectures/week05/09.2-functions.cpp
@@ -1,3 +1,3 @@
 #include "09.2-functions.h"
 
-long f1(long p1, long p2) { return p1 * p2; }
+int f1(int p1, int p2) { return p1 * p2; }
